License added: CC-BY-SA 3.0
To enable other people the usage of this text a public license was added.
Although it's generally better to use a software license like GPL for the
source code, it was decided to license the source code here under the same
license, because it's basically a demonstration code base and not really meant
for any productive use.

As the author of this commit and all appending changes I, Erik Bernoth, hereby
grant full CC-BY-SA 3.0 rights to all contributions added by this commit.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>